Fix. 좌표 int -> float

+++ api/schemas/map_sc.py
@@ -4,8 +4,8 @@ from pydantic import BaseModel
 # 먼 지도
 class AreaList(BaseModel):
     areaCode: int
-    lat: int
-    long: int
+    lat: float
+    long: float
     businessCount: int
 
 class MapFarSchema(BaseModel):
